Fix solaris building on solaris 11

+++ build/solaris/setup.sh
@@ -27,9 +27,9 @@ fi
 
 update_py_packages() {
     if [ "$ARCH" == "sparc" ] && [ $SOLARIS -eq 11 ]; then
-        $PYTHONBIN -m pip install -r  $BUILD_DIR/solaris/require.sparc.txt --upgrade
+        $PYTHONBIN -m pip install -r  $BUILD_DIR/solaris/require.solaris.txt --upgrade
     else
-        CPPFLAGS="-I$LIBFFI_DEV" LDFLAGS='-Wl,-rpath,\${ORIGIN} -Wl,-rpath,\${ORIGIN}/lib' $PYTHONBIN -m pip install -r $BUILD_DIR/resources/require.txt --upgrade --no-binary :all:
+        CPPFLAGS="-I$LIBFFI_DEV" LDFLAGS='-Wl,-rpath,\${ORIGIN} -Wl,-rpath,\${ORIGIN}/lib' $PYTHONBIN -m pip install -r $BUILD_DIR/resources/require.solaris.txt --upgrade --no-binary :all:
     fi
 }
 
@@ -45,7 +45,9 @@ install_prereqs() {
     if [ $SOLARIS -eq 10 ]; then
         pkgutil -y -i gcc5core python27 python27_dev py_pip wget libffi_dev libssl_dev        
     else
-        pkg install gcc libffi zlib
+        if [ ! -f /usr/bin/gcc ]; then
+            pkg install gcc libffi zlib
+        fi
     fi
